Attempt to Treat Coronavirus Disease (COVID-19) with Electromagnetic Wave Irradiation
The effectiveness of a combination of frequency, intensity, etc., in the natural vibration of microwaves targeting the collapse of the most delicate part, the Heptad Repeat 2 (HR 2) region, in spike proteins of severe acute respiratory syndrome coronavirus 2 (SARS-CoV-2) is investigated. Based on the results, inactivation of the virus is attempted through the use of irradiating microwaves directed toward the infected site from outside the body.